Skip this

+++ Assets/Scripts/Patient.cs
@@ -3,7 +3,7 @@ using System.Collections;
 
 public class Patient : Person {
 	// Levels of sickness: 1 is difficult, 5 is easy.
-	public double sickness = 1;
+	public int sickness = 1;
 
 	private bool isInBed;
 	public bool IsInBed {
